+++ src/resources/abstract.create.ts
@@ -116,7 +116,9 @@ export default abstract class AbstractCreate extends AbstractOperation {
     var body: string = '';
 
     for (let attr in definition) {
-      if (definition[attr].type == 'object' && definition[attr].properties) {
+      if(definition[attr].readOnly === true || attr === '_links') {
+        continue;
+      } else if (definition[attr].type == 'object' && definition[attr].properties) {
         body += ''.padStart(depth, ' ') + attr + ':\n';
 
         body += this.createTemplate(definition[attr].properties, depth + 2);